+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1420,3 +1420,30 @@ class EstimateNonResCapacityTrackerFilledPostsSchemas:
             IndCQC.ct_non_res_filled_post_estimate_source: pl.String,
         }
     )
+
+
+@dataclass
+class TestJoinEstimatesToAscwds:
+    TEST_ROLES = ["role_a", "role_b"]
+    estimates_schema = pl.Schema(
+        {
+            "id": pl.Int32,
+            IndCQC.ascwds_workplace_import_date: pl.String,
+            IndCQC.establishment_id: pl.String,
+        }
+    )
+    ascwds_schema = pl.Schema(
+        {
+            IndCQC.ascwds_workplace_import_date: pl.String,
+            IndCQC.establishment_id: pl.String,
+            IndCQC.main_job_role_clean_labelled: pl.Enum(TEST_ROLES),
+            "value": pl.Float64,
+        }
+    )
+    expected_schema = pl.Schema(
+        {
+            "id": pl.Int32,
+            IndCQC.main_job_role_clean_labelled: pl.Enum(TEST_ROLES),
+            "value": pl.Float64,
+        }
+    )