+++ CHANGELOG.md
@@ -54,6 +54,8 @@ All notable changes to this project will be documented in this file.
   Removed automatic copying of raw data files into branch.
   Removed a batch of files no longer used.
 
+- Refactored merge job within Estimates by Job Role Pipeline.
+
 ### Fixed
 - Refactored `remove_duplicates_based_on_column_order` to give the same outputs for every run.